NAS-3043: SDK - refactor catalog and compute
separate and split models
preparation for adding new objects and services
