+++ .flake8
@@ -11,5 +11,6 @@ extend-ignore = E203, W504
 # Ignoring unused imports in these as they are API re-exports
 per-file-ignores =
     gooddata-sdk/gooddata_sdk/__init__.py:F401
+    gooddata-sdk/gooddata_sdk/*/__init__.py:F401
     gooddata-pandas/gooddata_pandas/__init__.py:F401
     gooddata-fdw/gooddata_fdw/__init__.py:F401
