+++ gooddata-fdw/gooddata_fdw/filter.py
@@ -8,14 +8,7 @@ from typing import Union
 import gooddata_fdw.column_utils as column_utils
 from gooddata_fdw.environment import ColumnDefinition, Qual
 from gooddata_fdw.pg_logging import _log_debug, _log_info
-from gooddata_sdk.compute_model import (
-    AbsoluteDateFilter,
-    Attribute,
-    Filter,
-    NegativeAttributeFilter,
-    ObjId,
-    PositiveAttributeFilter,
-)
+from gooddata_sdk import AbsoluteDateFilter, Attribute, Filter, NegativeAttributeFilter, ObjId, PositiveAttributeFilter
 
 # Once AbsoluteDateFilter supports empty from/to, remove this workaround
 MIN_DATE = "0001-01-01"
