+++ gooddata-fdw/gooddata_fdw/import_workspace.py
@@ -14,10 +14,7 @@ from gooddata_fdw.naming import (
     InsightColumnNamingStrategy,
 )
 from gooddata_fdw.pg_logging import _log_debug, _log_info, _log_warn
-from gooddata_sdk.catalog import Catalog
-from gooddata_sdk.compute_model import ObjId
-from gooddata_sdk.insight import Insight, InsightAttribute, InsightMetric
-from gooddata_sdk.sdk import GoodDataSdk
+from gooddata_sdk import CatalogWorkspaceContent, GoodDataSdk, Insight, InsightAttribute, InsightMetric, ObjId
 
 
 def _metric_format_to_precision(metric_format: Optional[str]) -> Optional[str]:
@@ -86,7 +83,7 @@ class InsightsWorkspaceImporter(WorkspaceImporter):
             f"headers_host={self._server_options.headers_host}"
         )
         _log_debug("loading full catalog")
-        catalog = self._sdk.catalog.get_full_catalog(self._workspace)
+        catalog = self._sdk.catalog_workspace_content.get_full_catalog(self._workspace)
         _log_debug("loading all insights")
         insights = self._sdk.insights.get_insights(self._workspace)
 
@@ -111,7 +108,7 @@ class InsightsWorkspaceImporter(WorkspaceImporter):
 
         return tables
 
-    def _table_columns_from_insight(self, insight: Insight, catalog: Catalog) -> list[ColumnDefinition]:
+    def _table_columns_from_insight(self, insight: Insight, catalog: CatalogWorkspaceContent) -> list[ColumnDefinition]:
         column_naming = DefaultInsightColumnNaming()
         attr_cols = [self._attribute_to_table_column(attr, column_naming, catalog) for attr in insight.attributes]
         metric_cols = [self._metric_to_table_column(metric, column_naming, catalog) for metric in insight.metrics]
@@ -120,7 +117,7 @@ class InsightsWorkspaceImporter(WorkspaceImporter):
 
     @staticmethod
     def _attribute_to_table_column(
-        attr: InsightAttribute, column_naming: InsightColumnNamingStrategy, catalog: Catalog
+        attr: InsightAttribute, column_naming: InsightColumnNamingStrategy, catalog: CatalogWorkspaceContent
     ) -> ColumnDefinition:
         column_name = column_naming.col_name_for_attribute(attr)
         _log_debug(f'creating col def "{column_name}" for attribute "{attr.label_id}"')
@@ -135,7 +132,7 @@ class InsightsWorkspaceImporter(WorkspaceImporter):
         )
 
     def _metric_to_table_column(
-        self, metric: InsightMetric, column_naming: InsightColumnNamingStrategy, catalog: Catalog
+        self, metric: InsightMetric, column_naming: InsightColumnNamingStrategy, catalog: CatalogWorkspaceContent
     ) -> ColumnDefinition:
         column_name = column_naming.col_name_for_metric(metric)
         metric_format = self._get_insight_metric_format(metric, catalog)
@@ -159,7 +156,7 @@ class InsightsWorkspaceImporter(WorkspaceImporter):
 
     # InsightMetric do not contain format in case of stored metrics
     @staticmethod
-    def _get_insight_metric_format(metric: InsightMetric, catalog: Catalog) -> Optional[str]:
+    def _get_insight_metric_format(metric: InsightMetric, catalog: CatalogWorkspaceContent) -> Optional[str]:
         if metric.format:
             return metric.format
         elif metric.item_id:
@@ -184,7 +181,7 @@ class SemanticLayerWorkspaceImporter(WorkspaceImporter):
             f"headers_host={self._server_options.headers_host}"
         )
 
-        catalog = self._sdk.catalog.get_full_catalog(self._workspace)
+        catalog = self._sdk.catalog_workspace_content.get_full_catalog(self._workspace)
         naming = DefaultCatalogNamingStrategy()
         columns = []
