+++ gooddata-fdw/gooddata_fdw/result_reader.py
@@ -5,7 +5,7 @@ from typing import Any, Generator
 
 import gooddata_fdw.column_validation as col_valid
 from gooddata_fdw.environment import ColumnDefinition
-from gooddata_sdk.table import ExecutionTable
+from gooddata_sdk import ExecutionTable
 from gooddata_sdk.type_converter import DBTypeConverterStore
