+++ gooddata-sdk/gooddata_sdk/catalog/__init__.py
@@ -1,0 +1,1 @@
+# (C) 2022 GoodData Corporation
