+++ gooddata-sdk/gooddata_sdk/catalog/entity.py
@@ -1,0 +1,41 @@
+# (C) 2022 GoodData Corporation
+from __future__ import annotations
+
+from typing import Any, Optional
+
+from gooddata_sdk.compute.model.base import ObjId
+
+
+class CatalogEntity:
+    def __init__(self, entity: dict[str, Any]) -> None:
+        self._e = entity["attributes"]
+        self._entity = entity
+        self._obj_id = ObjId(self._entity["id"], type=self._entity["type"])
+
+    @property
+    def id(self) -> str:
+        return self._entity["id"]
+
+    @property
+    def type(self) -> str:
+        return self._entity["type"]
+
+    @property
+    def title(self) -> Optional[str]:
+        # Optional, not all metadata objects contain title
+        return self._e.get("title")
+
+    @property
+    def description(self) -> Optional[str]:
+        # Optional, not all metadata objects contain description
+        return self._e.get("description")
+
+    @property
+    def obj_id(self) -> ObjId:
+        return self._obj_id
+
+    def __str__(self) -> str:
+        return self.__repr__()
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}(id={self.id}, title={self.title})"
