+++ gooddata-sdk/gooddata_sdk/catalog/types.py
@@ -1,0 +1,7 @@
+# (C) 2022 GoodData Corporation
+from __future__ import annotations
+
+from typing import Dict, Set
+
+# Use typing collection types to support python < py3.9
+ValidObjects = Dict[str, Set[str]]
