+++ gooddata-sdk/gooddata_sdk/catalog/workspace/model/content_objects/__init__.py
@@ -1,0 +1,1 @@
+# (C) 2022 GoodData Corporation
