+++ gooddata-sdk/gooddata_sdk/catalog/workspace/service.py
@@ -1,0 +1,130 @@
+# (C) 2022 GoodData Corporation
+from __future__ import annotations
+
+import functools
+from typing import List, Union
+
+import gooddata_afm_client.apis as afm_apis
+import gooddata_afm_client.models as afm_models
+import gooddata_metadata_client.apis as metadata_apis
+from gooddata_sdk.catalog.types import ValidObjects
+from gooddata_sdk.catalog.workspace.model.content_objects.dataset import CatalogFact, CatalogLabel
+from gooddata_sdk.catalog.workspace.model.content_objects.metric import CatalogMetric
+from gooddata_sdk.catalog.workspace.model_container import CatalogWorkspaceContent
+from gooddata_sdk.client import GoodDataApiClient
+from gooddata_sdk.compute.model.attribute import Attribute
+from gooddata_sdk.compute.model.execution import ExecutionDefinition, compute_model_to_api_model
+from gooddata_sdk.compute.model.filter import Filter
+from gooddata_sdk.compute.model.metric import Metric
+from gooddata_sdk.utils import load_all_entities
+
+ValidObjectTypes = Union[Attribute, Metric, Filter, CatalogLabel, CatalogFact, CatalogMetric]
+
+# Use typing collection types to support python < py3.9
+ValidObjectsInputType = Union[ValidObjectTypes, List[ValidObjectTypes], ExecutionDefinition]
+
+
+class CatalogWorkspaceContentService:
+    # Note on the disabled checking:
+    # generated client has issues parsing the vis objects; .. have to avoid return type checks
+    #
+    # note: the parsing is done lazily so it does not necessarily bomb on the next line but when trying to
+    #  access returned object's properties
+
+    def __init__(self, api_client: GoodDataApiClient) -> None:
+        self._client = api_client
+        self._entities_api = metadata_apis.EntitiesApi(api_client.metadata_client)
+        self._afm_actions_api = afm_apis.ActionsApi(api_client.afm_client)
+
+    def get_full_catalog(self, workspace_id: str) -> CatalogWorkspaceContent:
+        """
+        Retrieves catalog for a workspace. Catalog contains all data sets and metrics defined in that workspace.
+
+        :param workspace_id: workspace identifier
+        :return:
+        """
+        get_datasets = functools.partial(
+            self._entities_api.get_all_entities_datasets,
+            workspace_id,
+            include=["attributes", "facts"],
+            _check_return_type=False,
+        )
+
+        get_attributes = functools.partial(
+            self._entities_api.get_all_entities_attributes,
+            workspace_id,
+            include=["labels"],
+            _check_return_type=False,
+        )
+
+        get_metrics = functools.partial(
+            self._entities_api.get_all_entities_metrics, workspace_id, _check_return_type=False
+        )
+
+        attributes = load_all_entities(get_attributes)
+        datasets = load_all_entities(get_datasets)
+        metrics = load_all_entities(get_metrics)
+
+        valid_obj_fun = functools.partial(self.compute_valid_objects, workspace_id)
+
+        return CatalogWorkspaceContent.create_workspace_content_catalog(valid_obj_fun, datasets, attributes, metrics)
+
+    @staticmethod
+    def _prepare_afm_for_availability(items: list[ValidObjectTypes]) -> afm_models.AFM:
+        attributes = []
+        metrics = []
+        filters = []
+
+        for item in items:
+            if isinstance(item, Attribute):
+                attributes.append(item)
+            elif isinstance(item, Metric):
+                metrics.append(item)
+            elif isinstance(item, Filter):
+                filters.append(item)
+            elif isinstance(item, CatalogLabel):
+                attributes.append(item.as_computable())
+            elif isinstance(item, (CatalogFact, CatalogMetric)):
+                metrics.append(item.as_computable())
+
+        return compute_model_to_api_model(attributes=attributes, metrics=metrics, filters=filters)
+
+    def compute_valid_objects(self, workspace_id: str, ctx: ValidObjectsInputType) -> ValidObjects:
+        """
+        Returns attributes, facts, and metrics which are valid to add to a context that already
+        contains some entities from the semantic model. The entities are typically used to compute analytics and
+        come from the execution definition. You may, however, specify the entities through different layers of
+        convenience.
+
+        :param workspace_id: workspace identifier
+        :param ctx: items already in context. you can specify context in one of the following ways:
+            - single item or list of items from the execution model
+            - single item or list of items from catalog model; catalog fact, label or metric may be added
+            - the entire execution definition that is used to compute analytics
+
+        :return: a dict of sets; type of available object is used as key in the dict,
+            the value is a set containing id's of available items
+        """
+        if isinstance(ctx, ExecutionDefinition):
+            afm = compute_model_to_api_model(attributes=ctx.attributes, metrics=ctx.metrics, filters=ctx.filters)
+        else:
+            _ctx = ctx if isinstance(ctx, list) else [ctx]
+            afm = self._prepare_afm_for_availability(_ctx)
+
+        query = afm_models.AfmValidObjectsQuery(afm=afm, types=["facts", "attributes", "measures"])
+        response = self._afm_actions_api.compute_valid_objects(workspace_id=workspace_id, afm_valid_objects_query=query)
+
+        by_type: dict[str, set[str]] = dict()
+
+        for available in response.items:
+            _type = available["type"]
+
+            if _type not in by_type:
+                items_of_type: set[str] = set()
+                by_type[_type] = items_of_type
+            else:
+                items_of_type = by_type[_type]
+
+            items_of_type.add(available["id"])
+
+        return by_type
