+++ gooddata-sdk/gooddata_sdk/type_converter.py
@@ -218,20 +218,20 @@ class ConverterRegistryStore:
             return cls._TYPE_REGISTRIES[type_name]
 
     @classmethod
-    def register(cls, type_name: str, class_: Type[Converter], sub_types: Optional[list[str]] = None) -> None:
+    def register(cls, type_name: str, class_converter: Type[Converter], sub_types: Optional[list[str]] = None) -> None:
         """
         Register Converter instance created from provided Converter class to given type and list of sub types.
         When sub types are not provided, converter is registered as the default one for given type.
         :param type_name: type name
-        :param class\\_: Converter class
+        :param class_converter: Converter class
         :param sub_types: list of sub types or None (default type Converter)
         """
         registry = cls._get_registry(type_name)
         if sub_types is None:
-            registry.register(class_(), None)
+            registry.register(class_converter(), None)
         else:
             for sub_type in sub_types:
-                registry.register(class_(), sub_type)
+                registry.register(class_converter(), sub_type)
 
     @classmethod
     def find_converter(cls, type_name: str, sub_type: Optional[str] = None) -> Converter:
