+++ gooddata-sdk/tests/catalog/test_catalog.py
@@ -17,7 +17,7 @@ gd_vcr = vcr.VCR(filter_headers=["authorization", "user-agent"], serializer="jso
 @gd_vcr.use_cassette(str(_fixtures_dir / "demo_catalog.json"))
 def test_catalog_load(test_config):
     sdk = GoodDataSdk.create(host_=test_config["host"], token_=test_config["token"])
-    catalog = sdk.catalog.get_full_catalog(test_config["workspace"])
+    catalog = sdk.catalog_workspace_content.get_full_catalog(test_config["workspace"])
 
     # rough initial smoke-test; just do a quick 'rub'
     assert len(catalog.metrics) == 24
@@ -33,7 +33,7 @@ def test_catalog_load(test_config):
 @gd_vcr.use_cassette(str(_fixtures_dir / "demo_catalog_availability.json"))
 def test_catalog_availability(test_config):
     sdk = GoodDataSdk.create(host_=test_config["host"], token_=test_config["token"])
-    catalog = sdk.catalog.get_full_catalog(test_config["workspace"])
+    catalog = sdk.catalog_workspace_content.get_full_catalog(test_config["workspace"])
     claim_count = catalog.get_metric("campaign_spend")
 
     filtered_catalog = catalog.catalog_with_valid_objects(claim_count)
